# Optimize 'family' query (330ms -> 16ms)