TL;DR: postgres cannot optimize the 3 queries that are OR'd together and so performs sequential scans. Splitting the queries into 3 separate ones and union-ing the results allows for the indexes to be used and thus no sequential scan, thus no ridiculous performance hit:

**BEFORE**
```
Gather Merge  (cost=86751.54..86752.01 rows=4 width=1040) (actual time=272.383..336.663 rows=3 loops=1)
  Workers Planned: 2
  Workers Launched: 2
  Buffers: shared hit=14289 read=49872, temp read=23480 written=23544
  ->  Sort  (cost=85751.52..85751.52 rows=2 width=1040) (actual time=268.480..268.484 rows=1 loops=3)
        Sort Key: coreapp_scratch.creation_time
        Sort Method: quicksort  Memory: 25kB
        Buffers: shared hit=14289 read=49872, temp read=23480 written=23544
        Worker 0:  Sort Method: quicksort  Memory: 28kB
        Worker 1:  Sort Method: quicksort  Memory: 25kB
        ->  Parallel Hash Join  (cost=23140.33..85751.51 rows=2 width=1040) (actual time=253.052..268.369 rows=1 loops=3)
              Hash Cond: ((coreapp_scratch.target_assembly_id)::text = (coreapp_assembly.hash)::text)
              Join Filter: (((coreapp_scratch.parent_id)::text = 'b50e1d94086caf49dd0a3acce3f4612a80a46d595ac982330a8185efd70ec4b8'::text) OR ((coreapp_scratch.family_id)::text = 'F60aW'::text) OR ((coreapp_assembly.source_asm_id)::text = 'WiZaF'::text))
              Rows Removed by Join Filter: 53906
              Buffers: shared hit=14175 read=49872, temp read=23480 written=23544
              ->  Parallel Seq Scan on coreapp_scratch  (cost=0.00..43866.62 rows=67262 width=1040) (actual time=0.026..63.240 rows=53907 loops=3)
                    Buffers: shared hit=12947 read=30247
              ->  Parallel Hash  (cost=21346.70..21346.70 rows=55970 width=130) (actual time=58.470..58.471 rows=44778 loops=3)
                    Buckets: 32768  Batches: 8  Memory Usage: 3072kB
                    Buffers: shared hit=1162 read=19625, temp written=2184
                    ->  Parallel Seq Scan on coreapp_assembly  (cost=0.00..21346.70 rows=55970 width=130) (actual time=0.029..32.881 rows=44778 loops=3)
                          Buffers: shared hit=1162 read=19625
Planning:
  Buffers: shared hit=17
Planning Time: 0.494 ms
Execution Time: 336.707 ms
```

**AFTER**
```
Sort  (cost=42.21..42.22 rows=4 width=2347) (actual time=16.046..16.050 rows=3 loops=1)
  Sort Key: coreapp_scratch.creation_time
  Sort Method: quicksort  Memory: 28kB
  ->  HashAggregate  (cost=42.13..42.17 rows=4 width=2347) (actual time=16.036..16.041 rows=3 loops=1)
        Group Key: coreapp_scratch.slug, coreapp_scratch.name, coreapp_scratch.description, coreapp_scratch.creation_time, coreapp_scratch.last_updated, coreapp_scratch.compiler, coreapp_scratch.platform, coreapp_scratch.compiler_flags, coreapp_scratch.diff_flags, coreapp_scratch.preset_id, coreapp_scratch.target_assembly_id, coreapp_scratch.source_code, coreapp_scratch.context, coreapp_scratch.diff_label, coreapp_scratch.score, coreapp_scratch.max_score, coreapp_scratch.match_override, coreapp_scratch.libraries, coreapp_scratch.family_id, coreapp_scratch.parent_id, coreapp_scratch.owner_id, coreapp_scratch.claim_token
        Batches: 1  Memory Usage: 48kB
        ->  Append  (cost=12.46..41.91 rows=4 width=2347) (actual time=0.665..1.787 rows=4 loops=1)
              ->  Sort  (cost=12.46..12.47 rows=2 width=1040) (actual time=0.665..0.666 rows=3 loops=1)
                    Sort Key: coreapp_scratch.creation_time DESC
                    Sort Method: quicksort  Memory: 28kB
                    ->  Index Scan using coreapp_scratch_family_id_a38ac945 on coreapp_scratch  (cost=0.42..12.45 rows=2 width=1040) (actual time=0.642..0.648 rows=3 loops=1)
                          Index Cond: ((family_id)::text = 'F60aW'::text)
              ->  Sort  (cost=8.45..8.45 rows=1 width=1040) (actual time=0.448..0.449 rows=1 loops=1)
                    Sort Key: coreapp_scratch_1.creation_time DESC
                    Sort Method: quicksort  Memory: 26kB
                    ->  Index Scan using coreapp_scratch_parent_id_1c2e2d85_like on coreapp_scratch coreapp_scratch_1  (cost=0.42..8.44 rows=1 width=1040) (actual time=0.440..0.442 rows=1 loops=1)
                          Index Cond: ((parent_id)::text = 'WiZaF'::text)
              ->  Sort  (cost=20.92..20.93 rows=1 width=1040) (actual time=0.666..0.667 rows=0 loops=1)
                    Sort Key: coreapp_scratch_2.creation_time DESC
                    Sort Method: quicksort  Memory: 25kB
                    ->  Nested Loop  (cost=0.84..20.91 rows=1 width=1040) (actual time=0.663..0.663 rows=0 loops=1)
                          ->  Index Scan using coreapp_assembly_source_asm_id_ec245923_like on coreapp_assembly  (cost=0.42..8.44 rows=1 width=65) (actual time=0.662..0.662 rows=0 loops=1)
                                Index Cond: ((source_asm_id)::text = 'b50e1d94086caf49dd0a3acce3f4612a80a46d595ac982330a8185efd70ec4b8'::text)
                          ->  Index Scan using coreapp_scratch_target_assembly_id_27f08bfe on coreapp_scratch coreapp_scratch_2  (cost=0.42..12.45 rows=2 width=1040) (never executed)
                                Index Cond: ((target_assembly_id)::text = (coreapp_assembly.hash)::text)
Planning Time: 0.538 ms
Execution Time: 16.115 ms
```